fix the budget from int to float

+++ src/main/java/com/medo/xbuilder/Dao/Project/ProjectDao.java
@@ -19,7 +19,7 @@ public class ProjectDao implements ProjectService {
             preparedStatement.setString(2,Project.getProjectDescription());
             preparedStatement.setDate(3, Project.getStartDate());
             preparedStatement.setDate(4, Project.getEndDate());
-            preparedStatement.setInt(5,Project.getBudget());
+            preparedStatement.setFloat(5,Project.getBudget());
             preparedStatement.executeUpdate();
             return true;
 
@@ -40,7 +40,7 @@ public class ProjectDao implements ProjectService {
             preparedStatement.setString(2,Project.getProjectDescription());
             preparedStatement.setDate(3,Project.getStartDate());
             preparedStatement.setDate(4,Project.getEndDate());
-            preparedStatement.setInt(5,Project.getBudget());
+            preparedStatement.setFloat(5,Project.getBudget());
             preparedStatement.setInt(6,Project.getProjectId());
                 return true;
